GUI Updates and MAVComms Tweaks
Added a function to change the frames on the fly, allowing different
missions to have different inputs or readouts. Added a .join() call to
the mavcomms start thread, to wait for the thread to exit with a
repsonse. Although this means the GUI freezes, it ensures an error free
connection (or ease of debugging on no connection). For missionPayload,
only change was checking velocity is greater than zero, rather than not
equal to zero.